# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemberCreate": {
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "temporary_password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Temporary Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "display_name",
          "temporary_password",
          "role"
        ],
        "title": "MemberCreate",
        "type": "object"
      },
      "MemberResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "status": {
            "$ref": "#/components/schemas/MembershipStatus"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "email",
          "display_name",
          "role",
          "status",
          "created_at"
        ],
        "title": "MemberResponse",
        "type": "object"
      },
      "MemberUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TenantRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "MemberUpdate",
        "type": "object"
      },
      "MembershipStatus": {
        "enum": [
          "active",
          "disabled"
        ],
        "title": "MembershipStatus",
        "type": "string"
      },
...
      "ModelActivateRequest": {
        "properties": {
          "application_id": {
            "format": "uuid",
            "title": "Application Id",
            "type": "string"
          }
        },
        "required": [
          "application_id"
        ],
        "title": "ModelActivateRequest",
        "type": "object"
      },
...
      "ModelConfigCreate": {
        "properties": {
          "embedding_dimension": {
            "anyOf": [
              {
                "maximum": 16384.0,
                "minimum": 8.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dimension"
          },
          "input_price_micros_per_million": {
            "default": 0,
            "minimum": 0.0,
            "title": "Input Price Micros Per Million",
            "type": "integer"
          },
          "max_tokens": {
            "default": 1024,
            "maximum": 128000.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "output_price_micros_per_million": {
            "default": 0,
            "minimum": 0.0,
            "title": "Output Price Micros Per Million",
            "type": "integer"
          },
          "provider_account_id": {
            "format": "uuid",
            "title": "Provider Account Id",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/ModelPurpose"
          },
          "temperature": {
            "default": 0.2,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "provider_account_id",
          "name",
          "model_name",
          "purpose"
        ],
        "title": "ModelConfigCreate",
        "type": "object"
      },
      "ModelConfigResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "embedding_dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Dimension"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_price_micros_per_million": {
            "title": "Input Price Micros Per Million",
            "type": "integer"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_price_micros_per_million": {
            "title": "Output Price Micros Per Million",
            "type": "integer"
          },
          "provider_account_id": {
            "format": "uuid",
            "title": "Provider Account Id",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/ModelPurpose"
          },
          "status": {
            "$ref": "#/components/schemas/ModelStatus"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "provider_account_id",
          "name",
          "model_name",
          "purpose",
          "embedding_dimension",
          "temperature",
          "max_tokens",
          "input_price_micros_per_million",
          "output_price_micros_per_million",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ModelConfigResponse",
        "type": "object"
      },
      "ModelPurpose": {
        "enum": [
          "chat",
          "embedding"
        ],
        "title": "ModelPurpose",
        "type": "string"
      },
      "ModelStatus": {
        "enum": [
          "inactive",
          "active"
        ],
        "title": "ModelStatus",
        "type": "string"
      },
...
      "ProviderAccountCreate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "ProviderAccountCreate",
        "type": "object"
      },
      "ProviderAccountResponse": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "has_api_key": {
            "title": "Has Api Key",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/ProviderScope"
          },
          "status": {
            "$ref": "#/components/schemas/ProviderStatus"
          },
          "tenant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "scope",
          "name",
          "kind",
          "base_url",
          "has_api_key",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProviderAccountResponse",
        "type": "object"
      },
      "ProviderKind": {
        "enum": [
          "fake",
          "openai_compatible"
        ],
        "title": "ProviderKind",
        "type": "string"
      },
      "ProviderScope": {
        "enum": [
          "platform",
          "tenant"
        ],
        "title": "ProviderScope",
        "type": "string"
      },
      "ProviderStatus": {
        "enum": [
          "draft",
          "ready",
          "disabled"
        ],
        "title": "ProviderStatus",
        "type": "string"
      },
      "ProviderTestResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProviderStatus"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "ProviderTestResponse",
        "type": "object"
      },
...
      "TenantRole": {
        "enum": [
          "tenant_admin",
          "agent"
        ],
        "title": "TenantRole",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AI Customer Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/ai/model-configs": {
      "get": {
        "operationId": "listModelConfigs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelConfigResponse"
                  },
                  "title": "Response Listmodelconfigs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Model Configs",
        "tags": [
          "model-gateway"
        ]
      },
      "post": {
        "operationId": "createModelConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Model Config",
        "tags": [
          "model-gateway"
        ]
      }
    },
    "/v1/admin/ai/model-configs/{model_config_id}/activate": {
      "post": {
        "operationId": "activateModelConfig",
        "parameters": [
          {
            "in": "path",
            "name": "model_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Config Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelActivateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activate Model Config",
        "tags": [
          "model-gateway"
        ]
      }
    },
    "/v1/admin/ai/model-configs/{model_config_id}/deactivate": {
      "post": {
        "operationId": "deactivateModelConfig",
        "parameters": [
          {
            "in": "path",
            "name": "model_config_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Config Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deactivate Model Config",
        "tags": [
          "model-gateway"
        ]
      }
    },
    "/v1/admin/ai/provider-accounts": {
      "get": {
        "operationId": "listProviderAccounts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderAccountResponse"
                  },
                  "title": "Response Listprovideraccounts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Provider Accounts",
        "tags": [
          "model-gateway"
        ]
      },
      "post": {
        "operationId": "createProviderAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderAccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Provider Account",
        "tags": [
          "model-gateway"
        ]
      }
    },
    "/v1/admin/ai/provider-accounts/{account_id}/test": {
      "post": {
        "operationId": "testProviderAccount",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Test Provider Account",
        "tags": [
          "model-gateway"
        ]
      }
    },
...
    "/v1/admin/members": {
      "get": {
        "operationId": "listTenantMembers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  },
                  "title": "Response Listtenantmembers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Members",
        "tags": [
          "tenant-members"
        ]
      },
      "post": {
        "operationId": "createTenantMember",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Member",
        "tags": [
          "tenant-members"
        ]
      }
    },
    "/v1/admin/members/{membership_id}": {
      "patch": {
        "operationId": "updateTenantMember",
        "parameters": [
          {
            "in": "path",
            "name": "membership_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Membership Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Member",
        "tags": [
          "tenant-members"
        ]
      }
    },
...
  }
}